+++ src/F2_autoreport.py
@@ -328,6 +328,7 @@ def insert_single_plot(
             template="simple_white",
             xaxis=styling_dict,
             yaxis=styling_dict,
+            font_family="sans-serif",
             title={
                 "text": plot_title,
                 "y": 0.90,
@@ -354,6 +355,7 @@ def insert_single_plot(
             xaxis_title=x_axis_name,
             yaxis_title=y_axis_name,
             template="simple_white",
+            font_family="sans-serif",
             xaxis=go.layout.XAxis(
                 showgrid=True,
                 gridwidth=1.5,
@@ -524,6 +526,7 @@ def insert_flows_plots(
     fig.update_layout(
         xaxis_title=x_legend,
         yaxis_title=y_legend,
+        font_family="sans-serif",
         template="simple_white",
         xaxis=styling_dict,
         yaxis=styling_dict,
@@ -535,9 +538,7 @@ def insert_flows_plots(
             "xanchor": "center",
             "yanchor": "top",
         },
-        legend=dict(
-            y=0.5, traceorder="normal", font=dict(family="sans-serif", color="black"),
-        ),
+        legend=dict(y=0.5, traceorder="normal", font=dict(color="black"),),
     )
 
     # Specific modifications for print version
@@ -627,6 +628,7 @@ def insert_pie_plots(
             "yanchor": "top",
             "pad": {"r": 5, "l": 5, "b": 5, "t": 5},
         },
+        font_family="sans-serif",
         height=500,
         width=700,
         autosize=True,
@@ -862,7 +864,7 @@ def create_app(results_json):
     # Creating a dataframe for the demands
     demands = results_json[ENERGY_CONSUMPTION]
 
-    ## Removing all columns that are not actually from demands
+    # Removing all columns that are not actually from demands
     drop_list = []
     for column_label in demands:
         # Identifies excess sink in demands for removal
@@ -874,7 +876,7 @@ def create_app(results_json):
         elif DSO_FEEDIN in column_label:
             drop_list.append(column_label)
 
-    # Remove droplist items (ie. sinks that are not demands) from data
+    # Remove some elements from drop_list (ie. sinks that are not demands) from data
     for item in drop_list:
         del demands[item]
 
@@ -1013,7 +1015,6 @@ def create_app(results_json):
 
     log_file = os.path.join(OUTPUT_FOLDER, "mvs_logfile.log")
     # log_file = "/home/mr/Projects/mvs_eland/MVS_outputs/mvs_logfile.log"
-    print(log_file)
 
     with open(log_file) as log_messages:
         log_messages = log_messages.readlines()