+++ CHANGELOG.md
@@ -25,6 +25,7 @@ Here is a template for new release sections
 ### Changed
  - Move and rename json converter and parser to B0 module (#464)
  - Modified json converter to avoid stringifying special types such as pandas.Dataframes (#464)
+ - Changed the font family used in the plots in F2_autoreport.py and changed the wording of some comments (#496)
 
 ### Removed
 -